Cleanup solution
