+++ framework/pages/notebookPage.js
@@ -6,9 +6,7 @@ const NotebookPage = function () {
     const clsErrorOutput = '.jp-Cell-outputWrapper [data-mime-type="application/vnd.jupyter.stderr"]';
     const divCodeCell = '.jp-Notebook .jp-CodeCell';
     const divCellIdx = '.jp-InputArea-prompt';
-    const tabNotebook = (fileName) => `li.lm-TabBar-tab[title*="Name: ${fileName}"]`;
-    //const tabNotebook = (fileName) => `.jp-DirListing-content`;
-    
+    const tabNotebook = (fileName) => `li.lm-TabBar-tab[title*="Name: ${fileName}"]`;    
 
     this.fileStatus = {
         Loaded: 'Formula Framework | Idle',
@@ -50,27 +48,13 @@ const NotebookPage = function () {
             async function (that) {
                 await page.waitForSelector(that.isFileLoaded(fileName));
             }
-        )
-
-        // let isBusy = '';
-        // try {
-        //     const busySelector = this.isFileBusy(fileName);
-        //     isBusy = await page.$eval(busySelector, el => {
-        //         return (el) ? el.innerText : '';
-        //     });
-        // }
-        // catch {};
-
-        // if (isBusy) {
-        //     await page.waitForSelector(this.isFileLoaded(fileName));
-        // };        
+        )      
     };
 
     this.getTabId = async function (page, fileName) {
         await page.waitForSelector(tabNotebook(fileName));
         let tab = await page.$(tabNotebook(fileName));
         let tabId = await page.evaluate(([tab]) => tab.getAttribute('data-id'), [tab]);
-        //console.log(tabNotebook(fileName) + ' TabId: ' +tabId);
         return tabId;
     };
 
@@ -86,19 +70,6 @@ const NotebookPage = function () {
                 await page.waitForSelector(btnDiscard);
                 await page.click(btnDiscard);
             });
-
-        // let isDialog = '';
-        // try {
-        //     isDialog = await page.$eval(btnDiscard, el => {
-        //         return (el) ? el.innerText : '';
-        //     });
-        // }
-        // catch {};
-
-        // if (isDialog) {
-        //     await page.waitForSelector(btnDiscard);
-        //     await page.click(btnDiscard);
-        // };
     };
 
     this.getCellIdx = async function(cell) {
@@ -107,19 +78,16 @@ const NotebookPage = function () {
 
     /**
      * 
-     * @param {ElementHandler} cell - runnable cell with code
+     * @param {ElementHandler} cell - runnable code cell
      * @returns - result of the cell run may contain computed result or an error message.
-     * intentional verification of the error class presense could result in exception, so it is wrapped with try-catch block
+     * intentional verification of the error class presense could result in exception,
+     * so it is wrapped with try-catch block
      * returns error message or empty
      */
     this.getCellError = async function(cell) {
         let isError = '';
 
         try {
-            // let input = await cell.$eval('.jp-Cell-inputWrapper', el => {
-            //     return (el) ? el.innerText : '';
-            // });
-            // console.log(input);
             isError = await cell.$eval(clsErrorOutput, el => {
                 return (el) ? el.innerText : '';
             });
